Cleaned up the code, eliminated compiler warnings
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@12347 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: e5bcaff034f36aafcc5d2d14f51701d4187e61b4

+++ vision/outlet_detection/src/main.cpp
@@ -245,7 +245,7 @@ int main(int argc,char** argv)
 			sprintf(filename, "%s/prec_rec.txt", output_path);
 			write_pr(filename, buf, outlet_roi, outlets);
 		}
-		printf(" found %d holes, %d outlets\n", features.size(), outlets.size());
+		printf(" found %d holes, %d outlets\n", (int)features.size(), (int)outlets.size());
 		
 		outletfarr2keypointarr(features, keypts);
